backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
openai
psycopg2-binary
dj-database-url
gunicorn
whitenoise
